cms-api: Extract dominant color computation into DamDominantColorService
Computing the dominant color of an uploaded image was the only part of FilesService
that used imgproxy, which made ImgproxyModule a hard dependency of the whole DAM
file handling.

Move the computation into DamDominantColorService, so ImgproxyService is injected in
one small service instead. FilesService.calculateDominantColor remains as a
deprecated delegator.
