+++ packages/api/cms-api/src/dam/files/files.service.spec.ts
@@ -38,9 +38,9 @@ function createServiceWithMockQueryBuilder() {
         null as never, // blobStorageBackendService
         null as never, // foldersService
         null as never, // DAM_CONFIG
-        null as never, // imgproxyService
         null as never, // orm
         null as never, // entityManager
+        null as never, // dominantColorService
     );
 
     const hasFolderConstraint = () =>
