+++ packages/api/cms-api/src/index.ts
@@ -192,6 +192,7 @@ export { createFolderEntity, FolderInterface } from "./dam/files/entities/folder
 export { FileImagesResolver } from "./dam/files/file-image.resolver";
 export { FilesService } from "./dam/files/files.service";
 export { FoldersService } from "./dam/files/folders.service";
+export { DamDominantColorService } from "./dam/images/dam-dominant-color.service";
 export { ImageInterface } from "./dam/images/dto/image.interface";
 export { HashImageParams, ImageParams } from "./dam/images/dto/image.params";
 export { ImageCropAreaInput } from "./dam/images/dto/image-crop-area.input";
